2.3.1.0 (#110)
* Update Dispatcher.php

* Update WarManager.php

* Update BattleRunner.php

* Update WarController.php

* Update actions.en.yml

* Update User.php

* Update Character.orm.xml

* Update settings.en.yml

* Update actions.en.yml

* Update messages.en.yml

* Update politics.en.yml

* Update places.en.yml

* Update communication.en.yml

* Update Dispatcher.php

* Update HouseManager.php

* Update CharacterManager.php

* Update ConversationManager.php

* Update CharacterSettingsType.php

* Create HouseSubcreateType.php

* Update HouseController.php

* Update GameRequestController.php

* Update create.html.twig

* Create subcreate.html.twig

* Update Dispatcher.php

* Update Dispatcher.php

* Update PlaceController.php

* Update PlacePermissionsSetType.php

* Update permissions.html.twig

* Update places.en.yml

* Update politics.en.yml

+++ src/BM2/SiteBundle/Service/WarManager.php
@@ -545,6 +545,9 @@ class WarManager {
 	}
 
 	public function disbandSiege(Siege $siege, Character $leader = null, $completed = FALSE) {
+		if ($siege->getBattle()) {
+			return false;
+		}
 		# Siege disbandment and removal actually happens as part of removeCharacterFromBattlegroup.
 		# This needs either completed to be true and leader to be null, or completed to be false and leader to be a Character.
 		$place = null;
@@ -678,6 +681,9 @@ class WarManager {
 	}
 
 	public function leaveSiege($character, $siege) {
+		if ($siege->getBattle()) {
+			return false;
+		}
 		foreach ($character->findActions('military.siege') as $action) {
 			#This should only ever be one, but just in case, and because findActions returns an ArrayCollection...
 			$this->em->remove($action);
